cong

+++ ShopT/ShopAuto/ShopAuto/Controllers/AccountsController.cs
@@ -5,6 +5,7 @@ using System.Drawing;
 using System.Drawing.Drawing2D;
 using System.IO;
 using System.Linq;
+using System.Net.Mail;
 using System.Web;
 using System.Web.Helpers;
 using System.Web.Mvc;
@@ -34,7 +35,7 @@ namespace ShopAuto.Controllers
             {
                 if (CheckEmail(user.Email))
                 {
-                    ViewBag.dangKy = "Email "+ user.Email + " đa tồn tại!";
+                    ViewBag.dangKy = "Email " + user.Email + " đa tồn tại!";
                     return View();
                 }
                 else
@@ -45,18 +46,18 @@ namespace ShopAuto.Controllers
                     user.Active = true;
                     _db.Users.Add(user);
                     _db.SaveChanges();
-                    ViewBag.dangKy = "Đăng ký thành công email "+user.Email;
+                    ViewBag.dangKy = "Đăng ký thành công email " + user.Email;
                     return View();
                 }
-                
+
             }
             ViewBag.dangKy = "Đăng ký thất bại!!";
             return View();
         }
         public bool CheckEmail(string email)
         {
-           
-            bool a = _db.Users.Where(n => n.Email.Contains(email)).ToList().Count !=0?true: false;
+
+            bool a = _db.Users.Where(n => n.Email.Contains(email)).ToList().Count != 0 ? true : false;
             return a;
         }
         public JsonResult SingIn(string acc, string pass)
@@ -82,7 +83,7 @@ namespace ShopAuto.Controllers
                 }
                 if (us != null && us.Active == false)
                 {
-                    query =2;
+                    query = 2;
                 }
             }
             return Json(query, JsonRequestBehavior.AllowGet);
@@ -97,7 +98,89 @@ namespace ShopAuto.Controllers
             Session.Remove("Id");
             return Redirect("/");
         }
+        public JsonResult SendCodeResetPass(string email)
+        {
+            var user = _db.Users.FirstOrDefault(n => n.Email == email);
+            if (user == null)
+            {
+                return Json(1, JsonRequestBehavior.AllowGet);
+            }
+            else
+            {
+                try
+                {
+                    int Numrd;
+                    var rd = new Random();
+                    Numrd = rd.Next(10000, 99999);
+                    user.CodeReset = Numrd;
+                    _db.Entry(user).State = EntityState.Modified;
+                    _db.SaveChanges();
+                    var shop = _db.ShopInformations.FirstOrDefault();
+                    System.Text.StringBuilder Body = new System.Text.StringBuilder();
+                    Body.Append("<p>Kính gửi anh(chị), " + shop.NameShop + " Xin chào quý khách </p>");
+                    Body.Append("<p>Chúng tôi đã nhận được yêu cầu đổi lại mật khẩu của anh(chị) vào lúc " + DateTime.Now + "</p>");
+                    Body.Append("<p>Chúng tôi xin gửi mã xác nhận : <span colspan='2'>" + Numrd + "</span></p>");
+                    Body.Append("<p>Chân thành chúc sức khỏe và cảm ơn bạn đã đọc.</p>");
+                    MailMessage mail = new MailMessage();
+                    mail.To.Add(email);
+                    mail.From = new MailAddress(shop.Email);
+                    mail.Subject = "Mã xác nhận.";//tiêu đề mail
+                    mail.Body = Body.ToString();// phần thân của mail ở trên
+                    mail.IsBodyHtml = true;
+                    SmtpClient smtp = new SmtpClient();
+                    smtp.Host = "smtp.gmail.com";
+                    //smtp.Port = 587;
+                    //smtp.UseDefaultCredentials = true;
+                    smtp.Credentials = new System.Net.NetworkCredential(shop.Email, shop.PassWordEmail);// tài khoản Gmail người gửi
+                    smtp.EnableSsl = true;
+                    smtp.Send(mail);
+                    return Json(2, JsonRequestBehavior.AllowGet);
+                }
+                catch (Exception)
+                {
+                    return Json(3, JsonRequestBehavior.AllowGet);
+                }
+            }
+        }
+        public JsonResult checkCode(string email, string code)
+        {
+            var user = _db.Users.FirstOrDefault(n => n.Email == email);
+            if (user == null)
+            {
+                return Json(1, JsonRequestBehavior.AllowGet);
+            }
+            if (user.CodeReset == Convert.ToInt32(code))
+            {
+                return Json(2, JsonRequestBehavior.AllowGet);
+            }
+            else { return Json(3, JsonRequestBehavior.AllowGet); }
+        }
+        public JsonResult ResetPass(string email, string pass)
+        {
+            var user = _db.Users.FirstOrDefault(n => n.Email == email);
+            if (user == null)
+            {
+                return Json(false, JsonRequestBehavior.AllowGet);
+            }
+            else
+            {
+                try
+                {
+                    string d = Encryptor.MD5Hash(pass);
+                    user.Password = d;
+                    user.CodeReset = 0;
+                    _db.Entry(user).State = EntityState.Modified;
+                    _db.SaveChanges();
+                    return Json(true, JsonRequestBehavior.AllowGet);
 
+                }
+                catch (Exception)
+                {
+                    return Json(false, JsonRequestBehavior.AllowGet);
+
+                }
+            }
+        }
         public User UserByEmail(string email)
         {
             return _db.Users.First(n => n.Email == email);
@@ -110,10 +193,21 @@ namespace ShopAuto.Controllers
         {
             if (Session["CusEmail"] == null)
             {
-
                 return RedirectToAction("redirectTo404");
             }
             return View(UserByEmail(Session["CusEmail"].ToString()));
+        }
+        public JsonResult UserLogin()
+        {
+            if (Session["Id"] != null)
+            {
+                int id = Convert.ToInt32(Session["Id"].ToString());
+                return Json(_db.Users.FirstOrDefault(n => n.ID == id), JsonRequestBehavior.AllowGet);
+            }
+            else
+            {
+                return Json(false, JsonRequestBehavior.AllowGet);
+            }
         }
         [HttpPost]
         public ActionResult Upload(HttpPostedFileBase image)
@@ -142,7 +236,7 @@ namespace ShopAuto.Controllers
                         user = UserByEmail(Session["CusEmail"].ToString());
                         //xoas hinh cu
                         FileInfo file = new FileInfo(Server.MapPath("~" + user.Image));
-                        if (file.Exists)
+                        if (file.Exists && !user.Image.Contains("hinhdd.jpg"))
                         {
                             file.Delete();
                         }
@@ -151,12 +245,12 @@ namespace ShopAuto.Controllers
                         user.Image = "/Content/img/TaiKhoan/" + fileName;
                         _db.Entry(user).State = EntityState.Modified;
                         _db.SaveChanges();
-                }
+                    }
                     catch (Exception ex)
-                {
-                    ViewBag.Message = "Lỗi:" + ex.Message.ToString();
+                    {
+                        ViewBag.Message = "Lỗi:" + ex.Message.ToString();
+                    }
                 }
-            }
                 else
                 {
                     ViewBag.Message = "Tập tin hình ảnh của bạn sai!";
@@ -172,7 +266,7 @@ namespace ShopAuto.Controllers
 
         }
         [HttpPost]
-        public ActionResult UpdateInformation( FormCollection f)
+        public ActionResult UpdateInformation(FormCollection f)
         {
             if (Session["IdCus"] == null)
             {
@@ -180,10 +274,10 @@ namespace ShopAuto.Controllers
             }
             else
             {
-                User account= UserByEmail(Session["CusEmail"].ToString());
+                User account = UserByEmail(Session["CusEmail"].ToString());
                 account.FirstName = f["FirstName"].ToString();
                 account.LastName = f["LastName"].ToString();
-                account.Phone = f["Phone"].ToString(); 
+                account.Phone = f["Phone"].ToString();
                 account.Addess = f["Address"].ToString();
                 _db.Entry(account).State = EntityState.Modified;
                 _db.SaveChanges();
@@ -203,14 +297,14 @@ namespace ShopAuto.Controllers
                 User account = UserByEmail(Session["CusEmail"].ToString());
                 string d = Encryptor.MD5Hash(account.Password);
                 string c = account.Password;
-                if (f["NewPass"].ToString()== f["NewPass2"].ToString()&&account.Password.Contains(Encryptor.MD5Hash(f["OldPass"].ToString())))
+                if (f["NewPass"].ToString() == f["NewPass2"].ToString() && account.Password.Contains(Encryptor.MD5Hash(f["OldPass"].ToString())))
                 {
-                    
+
                     account.Password = Encryptor.MD5Hash(f["NewPass"].ToString());
                     _db.Entry(account).State = EntityState.Modified;
                     _db.SaveChanges();
                 }
-                if(f["NewPass"].ToString() != f["NewPass2"].ToString())
+                if (f["NewPass"].ToString() != f["NewPass2"].ToString())
                 {
                     ViewBag.message2 = "Xác nhận mật khẩu mới sai!";
                 }
@@ -227,14 +321,14 @@ namespace ShopAuto.Controllers
             if (Session["IdCus"] == null)
             {
                 return RedirectToAction("AccountInformation");
-                
+
             }
             else
             {
-                id  = Convert.ToInt32(Session["IdCus"]);
+                id = Convert.ToInt32(Session["IdCus"]);
             }
-            return View(_db.Bills.Where(n=>n.Creator==id).OrderByDescending(n=>n.CreateDate).ToList());
+            return View(_db.Bills.Where(n => n.Creator == id).OrderByDescending(n => n.CreateDate).ToList());
         }
     }
-        
+
 }
